refactor: split large files, remove Keycodes.swift, clean up SwiftLint config
- Split MVClockView.swift into MVClockView.swift + MVClockView+Behavior.swift
  to comply with default file_length threshold
- Split TimerUITests.swift into 7 focused test files with shared base class
  to comply with default file_length and type_body_length thresholds
- Replace virtual keycode constants with event.charactersIgnoringModifiers,
  eliminating Keycodes.swift entirely (also improves international keyboard support)
- Remove SwiftLint excluded section and disabled_rules
- Enable number_separator opt-in rule
- Extract AppDelegate notification observers into observeNotifications()
