+++ .swiftlint.yml
@@ -1,9 +1,3 @@
-# rule identifiers to exclude from running
-disabled_rules:
-  - type_body_length
-  - function_body_length
-  - file_length
-
 # some rules are only opt-in
 # Find all the available rules by running: swiftlint rules
 opt_in_rules:
@@ -20,7 +14,7 @@ opt_in_rules:
   - joined_default_parameter
   - literal_expression_end_indentation
   - modifier_order
-  # - number_separator
+  - number_separator
   - operator_usage_whitespace
   - optional_enum_case_matching
   - prefer_self_type_over_type_of_self
@@ -46,8 +40,7 @@ included:
   - TimerTests
   - TimerUITests
 
-excluded:
-  - Timer/Keycodes.swift
+
 
 # reporter type (xcode, json, csv, checkstyle, junit, html, emoji, sonarqube, markdown)
 reporter: 'xcode'
