+++ TimerUITests/TimerUITestCase.swift
@@ -1,0 +1,16 @@
+import XCTest
+
+class TimerUITestCase: XCTestCase {
+  var app: XCUIApplication!
+
+  override func setUpWithError() throws {
+    continueAfterFailure = false
+    app = XCUIApplication()
+    app.launch()
+  }
+
+  override func tearDownWithError() throws {
+    app.terminate()
+    app = nil
+  }
+}
